fixed a bug with mutex file deletion
fixed bug with unitcode deleted object table fetching thing.

+++ django_snailtracker/helpers.py
@@ -75,7 +75,8 @@ def mutex_lock(key):
     try:
         yield True
     finally:
-        os.remove(mutex_key)
+        if os.path.exists(mutex_key):
+            os.remove(mutex_key)
 
 
 class SnailtrackerMutexLockedError(Exception):
